Require faraday 0.9.x
Any version of faraday less than 0.9.0 (at least up to 0.8.8) has a
conflict with the default_faraday_config method in configuration.rb.

The issue was that faraday used to use a simple Hash to hold options,
which does not respond to the #timeout= method. It was replaced with
the Options module's #connection_config which handles configuration
in a much more sane manner.